+++ java/org/lwjgl/opengl/GL15.java
@@ -522,7 +522,7 @@ public final class GL15 {
 	 */
 	public static ByteBuffer glMapBuffer(int target, int access) {
 		long __result = nglMapBuffer(target, access);
-		return memByteBuffer(__result, glGetBufferParameteri(target, GL_BUFFER_SIZE));
+		return memByteBuffer(__result, (int)glGetBufferParameteri(target, GL_BUFFER_SIZE));
 	}
 
 	/** Alternative version of: {@link #glMapBuffer MapBuffer} */
@@ -535,7 +535,7 @@ public final class GL15 {
 	/** Explicit size alternative version of: {@link #glMapBuffer MapBuffer} */
 	public static ByteBuffer glMapBuffer(int target, int access, long length, ByteBuffer old_buffer) {
 		long __result = nglMapBuffer(target, access);
-		return old_buffer != null && __result == memAddress0(old_buffer) && old_buffer.capacity() == length ? old_buffer : memByteBuffer(__result, length);
+		return old_buffer != null && __result == memAddress0(old_buffer) && old_buffer.capacity() == length ? old_buffer : memByteBuffer(__result, (int)length);
 	}
 
 	// --- [ glUnmapBuffer ] ---
